Create routes for registeration and home page
